Allowed passing a json to request a token. Created http.utils
